tests/wpt/tests/html/canvas/offscreen/fill-and-stroke-styles/2d.gradient.object.type.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.gradient.object.type</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.gradient.object.type</h1>
<p class="desc">window.CanvasGradient exists and has the right properties</p>

<p class="notes">Defined in "Web IDL" (draft)
<script>
var t = async_test("window.CanvasGradient exists and has the right properties");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  _assertDifferent(window.CanvasGradient, undefined, "window.CanvasGradient", "undefined");
  _assertDifferent(window.CanvasGradient.prototype.addColorStop, undefined, "window.CanvasGradient.prototype.addColorStop", "undefined");
  t.done();

});
</script>
